Overriding issue method to return random token
Overiging issue() method because by default it's calling the issue method in the AbstractAuthorizationGrantHandler class. That issue method need an actual user in the user store to properly work. Since MobileGrant use a tempory user as the authenticated user, the issue() method in the AbstractAuthorizationGrantHandler won't work properly. Since this is a sample application we can just return random id for tokens.